+++ src/collective/easynewsletter_combined_send/issuedatafetcher.py
@@ -42,10 +42,13 @@ class CombindSendDXIssueDataFetcher(DefaultDXIssueDataFetcher):
         content_parts = part_soup.select("#emailBody")
         anker_link_ref = "#lang_{0}".format(lang)
         anker_tag_name = "lang_{0}".format(lang)
+        anker_link_wrapper = output_soup.new_tag("p")
+        anker_link_wrapper.string = "> "
         anker_link = output_soup.new_tag("a", href=anker_link_ref)
-        anker_link.string = "> english version below"
+        anker_link.string = "english version below"
         anker_link["class"] = "english_version_below_link"
-        output_soup.select(".enlHeaderContent")[0].insert(0, anker_link)
+        anker_link_wrapper.append(anker_link)
+        output_soup.select(".enlHeaderContent")[0].insert(0, anker_link_wrapper)
         anker_tag = output_soup.new_tag("a")
         anker_tag["name"] = anker_tag_name
         output_soup.select(".aggregatedContentSlot")[0].append(anker_tag)